fix: dedupe role-formset JS, guard remove clicks, keep DELETE checkbox as no-JS fallback
