+++ teamshifts/templates/teamshifts/shift_create.html
@@ -95,12 +95,8 @@
                                     {% endif %}
                                 </td>
                                 <td>
-                                    {% if form.instance.pk %}
-                                        {{ form.DELETE }}
-                                    {% else %}
-                                        <button type="button" class="btn btn-danger btn-sm remove-role-btn"><i class="fa fa-trash"></i></button>
-                                        <span class="delete-checkbox hidden">{{ form.DELETE }}</span>
-                                    {% endif %}
+                                    <button type="button" class="btn btn-danger btn-sm remove-role-btn"><i class="fa fa-trash"></i></button>
+                                    <span class="delete-checkbox">{{ form.DELETE }}</span>
                                 </td>
                             </tr>
                         {% endfor %}
