+++ teamshifts/templates/teamshifts/shift_edit.html
@@ -75,7 +75,7 @@
                                 </td>
                                 <td>
                                     <button type="button" class="btn btn-danger btn-sm remove-role-btn"><i class="fa fa-trash"></i></button>
-                                    <span class="delete-checkbox hidden">{{ form.DELETE }}</span>
+                                    <span class="delete-checkbox">{{ form.DELETE }}</span>
                                 </td>
                             </tr>
                         {% endfor %}
